chore: adding hooks on write articles
